Move Provider to Prism/Providers + update docs
