+++ docs/advanced/custom-providers.md
@@ -4,27 +4,13 @@ Want to add support for a new AI provider in Prism? This guide will walk you thr
 
 ## Provider Interface
 
-All providers must implement the `Prism\Prism\Contracts\Provider` interface:
+All providers must extend the `Prism\Prism\Providers\Provider` class.
 
-```php
-use Prism\Prism\Embeddings\Request as EmbeddingsRequest;
-use Prism\Prism\Embeddings\Response as EmbeddingsResponse;
-use Prism\Prism\Structured\Request as StructuredRequest;
-use Prism\Prism\Structured\Response as StructuredResponse;
-use Prism\Prism\Text\Request as TextRequest;
-use Prism\Prism\Text\Response as TextResponse;
+The abstract class has a default method for all current required methods, though this may change.
 
-interface Provider
-{
-    public function text(TextRequest $request): TextResponse;
-
-    public function structured(StructuredRequest $request): StructuredResponse;
-
-    public function embeddings(EmbeddingsRequest $request): EmbeddingsResponse;
-}
-```
-
-The interface is intentionally simple, requiring just one method to handle text generation requests.
+Each provider should:
+- Overwrite the methods for the actions it supports. 
+- Overwrite the `handleRequestExceptions` method if supports advanced request exceptions.
 
 ## Registration Process
