+++ src/Providers/VoyageAI/VoyageAI.php
@@ -5,9 +5,9 @@ namespace Prism\Prism\Providers\VoyageAI;
 use Illuminate\Http\Client\PendingRequest;
 use Prism\Prism\Concerns\HandlesRequestExceptions;
 use Prism\Prism\Concerns\InitializesClient;
-use Prism\Prism\Contracts\Provider;
 use Prism\Prism\Embeddings\Request as EmbeddingRequest;
 use Prism\Prism\Embeddings\Response as EmbeddingsResponse;
+use Prism\Prism\Providers\Provider;
 
 class VoyageAI extends Provider
 {
