+++ src/Providers/XAI/XAI.php
@@ -7,7 +7,7 @@ namespace Prism\Prism\Providers\XAI;
 use Illuminate\Http\Client\PendingRequest;
 use Prism\Prism\Concerns\HandlesRequestExceptions;
 use Prism\Prism\Concerns\InitializesClient;
-use Prism\Prism\Contracts\Provider;
+use Prism\Prism\Providers\Provider;
 use Prism\Prism\Providers\XAI\Handlers\Text;
 use Prism\Prism\Text\Request as TextRequest;
 use Prism\Prism\Text\Response as TextResponse;
